Refactoring settings to use shared DTO and Enum between Core and Highlighter, adding proper settings data validation in frontend and backend, fixing compare renderings scripts to use updated CLI

+++ Calcpad.Core/Plotter/MapPlotter.cs
@@ -108,9 +108,9 @@ namespace Calcpad.Core
 
         private void GetImageSize()
         {
-            var w = (int)(Parser.GetSettingsVariable("PlotWidth", Settings.Width) * ScreenScaleFactor);
-            var h = (int)(Parser.GetSettingsVariable("PlotHeight", Settings.Height) * ScreenScaleFactor);
-            _size = (int)(Parser.GetSettingsVariable("PlotStep", Settings.Step) * ScreenScaleFactor);
+            var w = (int)((int)Parser.GetSettingsVariable("PlotWidth", Settings.Width) * ScreenScaleFactor);
+            var h = (int)((int)Parser.GetSettingsVariable("PlotHeight", Settings.Height) * ScreenScaleFactor);
+            _size = (int)((int)Parser.GetSettingsVariable("PlotStep", Settings.Step) * ScreenScaleFactor);
             var d = (int)Math.Sqrt(w * w + h * h);
             if (_size == 0)
             {
